cache tests

+++ tripy/tests/frontend/trace/ops/test_storage.py
@@ -78,7 +78,7 @@ class TestStorage:
     def test_empty_list(self):
         data = [[]]
         storage = Storage([], [TraceTensor("test", None, None, None, None, None)], data)
-        assert storage.dtype == tp.int32
+        assert storage.dtype == tp.float32
         assert storage.shape == (1, 0)
         assert storage.device.kind == "gpu"
         assert storage.data_str == "[[]]"
